ASYNC: add State::Preempted, and remove State::InQueue
Because more than 1 thread will invoke a task in parallel,
a new state of State::Preempted is introduced.

A task runner uses try_lock() when executing a task to
acquire the main routine.
Failing try_lock() means that the task is running on another thread
at that time, so, the task runner avoids to execute it and re-schedules.

Signed-off-by: Yuuki Takano <ytakanoster@gmail.com>